pyproject.toml
[tool.poetry]
name = "djangosample"
version = "0.1.0"
description = ""
authors = ["Naury <koertyf@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
redis = "^5.2.0"
punq = "^0.7.0"
pika = "^1.3.2"
pydantic-settings = "^2.6.1"
aiogram = "^3.15.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
